moved css files into styles folder and updated import paths
