+++ src/components/animal/AnimalCard.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from "react-router-dom"
-import './Animal.css';
+import './styles/Animal.css';
 
 const AnimalCard = props => {
     return (
