+++ src/components/animal/styles/Animal.css
