+++ src/components/animal/styles/AnimalDetail.css
